backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-cleanup==8.0.0
Pillow==10.4.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
whitenoise==6.5.0
python-dotenv==1.0.0
argon2-cffi==23.1.0
django-ratelimit==4.1.0
django-simple-captcha==0.6.0
django-redis==5.4.0
django-storages[boto3]==1.14.4